refactor(observability): session_id tagging moves to telemetry/session_tag.rs
telemetry.rs and its tests sat near the 600-line cap; the session-tag layer
is a self-contained seam (visitor, validator, layer, lookup) with its own
tests, so it becomes a sibling module. No behavior change.
